# More efficient implementation of `Lens` and `Prism` 