Below are results before and after with Scala 2.13 and JDK 21 on Intel® Core™ i7-11800H.

Before:
```
[info] Benchmark                    Mode  Cnt          Score            Error  Units
[info] LensGetBenchmark.direct     thrpt    5  722169232.835 ± 1015384026.194  ops/s
[info] LensGetBenchmark.monocle    thrpt    5   76777139.817 ±     756661.504  ops/s
[info] LensGetBenchmark.zioBlocks  thrpt    5    5698191.253 ±     224336.537  ops/s
[info] LensSetBenchmark.direct     thrpt    5  144129150.126 ±   11309434.405  ops/s
[info] LensSetBenchmark.monocle    thrpt    5   49347723.472 ±    2029335.400  ops/s
[info] LensSetBenchmark.quicklens  thrpt    5   19433522.273 ±    1113320.606  ops/s
[info] LensSetBenchmark.zioBlocks  thrpt    5    6162712.317 ±     112223.492  ops/s
```

After:
```
[info] Benchmark                    Mode  Cnt          Score         Error  Units
[info] LensGetBenchmark.direct     thrpt    5  948770241.627 ± 7166728.417  ops/s
[info] LensGetBenchmark.monocle    thrpt    5   76545914.567 ± 3157526.553  ops/s
[info] LensGetBenchmark.zioBlocks  thrpt    5   18837319.327 ± 4644917.196  ops/s
[info] LensSetBenchmark.direct     thrpt    5  144993599.231 ± 4827319.229  ops/s
[info] LensSetBenchmark.monocle    thrpt    5   47689009.852 ± 9510215.180  ops/s
[info] LensSetBenchmark.quicklens  thrpt    5   19631931.250 ±  560555.300  ops/s
[info] LensSetBenchmark.zioBlocks  thrpt    5   12165444.413 ± 1347561.989  ops/s
```